royalty: make elf robes, ñoldorin-elf armor/helmet and elf crown into noble apparel to satisfy the needs of noble title holders (robes only for lower tiers, the armor for any player pawns and crown only for highest ranks, same as vanilla crown) left the "apparel" tag in there to stay compatible if Royalty DLC is not installed.
Open issue: While the robe does satisfy the need, it is not listed in the negative thought a pawn gets when it's missing... So players might not immediately notice they can use them.
Not sure why, the same works fine for armor in the dwarves mod, and the elven headgear shows up just fine as well.