Lorien Deer Tweaks
Adding the Lorien Deer to Vanilla and modded Biomes. Commonality I put on the same biomes as vanilla deers, but way less commo, the idea being that while they prefer to live in the mallorn forests, they still have to migrate though the normal forests to reach other mallorn biomes.